requirements.txt
torch
torchvision
wandb
nltk
pandas
numpy
pillow
matplotlib
tqdm
scipy